move AuthenticationViewState to correct package (presentation) since its UI related
