+++ app/src/main/java/com/example/tasky/feature_login/presentation/AuthenticationViewState.kt
@@ -1,4 +1,4 @@
-package com.example.tasky.feature_login.domain.model
+package com.example.tasky.feature_login.presentation
 
 import com.example.tasky.common.data.util.DataError
